Move 'Note' model to 'base' package. Rename 'unimportant_notes.models.Note' to 'unimportant_notes.models.UnimportantNote'.
