+++ base/models.py
@@ -1,5 +1,7 @@
 from django.db import models
 
+from config.settings import AUTH_USER_MODEL
+
 
 class CreatedUpdatedBase(models.Model):
     """
@@ -20,3 +22,43 @@ class CreatedUpdatedBase(models.Model):
 
     class Meta:
         abstract = True
+
+
+class Note(CreatedUpdatedBase):
+    """
+    A note.
+    """
+
+    title = models.CharField(
+        "Title",
+        max_length=255,
+        help_text="The title of this note.",
+    )
+    author = models.ForeignKey(
+        AUTH_USER_MODEL,
+        on_delete=models.CASCADE,
+        related_name="notes",
+    )
+    content = models.TextField(
+        "Content",
+        help_text="The content of this note.",
+    )
+
+    def display_content(self):
+        """
+        This function returns a truncated version of the note's content.
+        This can be used in the admin panel and other places where the full
+        content is not needed.
+        """
+        return self.content[:50] + ("..." if len(self.content) > 50 else "")
+
+    def __str__(self):
+        return (
+            f"{self.title} - {self.content[:50]}"
+            f"{'...' if len(self.content) > 50 else ''}"
+        )
+
+    class Meta:
+        abstract = True
+        # `verbose_name` and `verbose_name_plural` and `ordering` are declared in the
+        # child class.
