+++ unimportant_notes/models.py
@@ -1,41 +1,13 @@
-from django.db import models
-
-from base.models import CreatedUpdatedBase
-from config.settings import AUTH_USER_MODEL
+from base.models import Note
 
 
-class Note(CreatedUpdatedBase):
+class UnimportantNote(Note):
     """
-    A note.
+    Model for `UnimportantNote`.
     """
-
-    title = models.CharField(
-        "Title",
-        max_length=255,
-        help_text="The title of this note.",
-    )
-    author = models.ForeignKey(
-        AUTH_USER_MODEL,
-        on_delete=models.CASCADE,
-        related_name="notes",
-    )
-    content = models.TextField(
-        "Content",
-        help_text="The content of this note.",
-    )
-
-    def display_content(self):
-        """
-        This function returns a truncated version of the note's content.
-        This can be used in the admin panel and other places where the full
-        content is not needed.
-        """
-        return self.content[:50] + ("..." if len(self.content) > 50 else "")
-
-    def __str__(self):
-        return f"{self.title}\n{self.content}"
 
     class Meta:
-        verbose_name = "Note"
-        verbose_name_plural = "Notes"
+        # Other `Meta` options are inherited from `Note`.
+        verbose_name = "Unimportant Note"
+        verbose_name_plural = "Unimportant Notes"
         ordering = ("-created",)
