+++ unimportant_notes/urls.py
@@ -4,6 +4,6 @@ from . import views
 
 app_name = "unimportant_notes"
 urlpatterns = [
-    path("notes/", views.NoteListView.as_view(), name="note_list"),
-    path("notes/create/", views.NoteCreateView.as_view(), name="note_create"),
+    path("", views.NoteListView.as_view(), name="note_list"),
+    path("create/", views.NoteCreateView.as_view(), name="note_create"),
 ]
